basics are working
